Update src.ino
You forgot to include the code that checks to see if the max pulse length timer has reset the 'p' flag before incrementing the pulse count in the ISR, so the debouncing wasn't actually working. This is also true of most of the pulse counting firmwares. With this modification even a nasty noisy reed switch counts correctly, whereas without it even using a RC filter the count was unreliable